Optimize least squares/two lines best fit
Replate numpy least squares with a more efficient implementation that takes advantage of the fact that the same squares are used multiple times in the two lines best fit algorithm. After the setup is done to pre-calculate the squares, the cost to caclulate a specific least squares solution is O(1). This has resulted in about an order of magnitude performance improvement (100-200ms to 10-20ms).

Now the largest delay is caused by buffering and polling for sensor data and is in the order of 50ms to 200ms.

Signed-off-by: Gareth Farrington <gareth@waves.ky>